# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for RiceSafe Application",
        "title": "RiceSafe API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/diseases": {
            "get": {
                "description": "Get a list of all diseases in the library",
                "tags": [
                    "Diseases"
                ],
                "summary": "List all diseases",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by category",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_disease.Disease"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new disease entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diseases"
                ],
                "summary": "Create a new disease",
                "parameters": [
                    {
                        "description": "Disease Data",
                        "name": "disease",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_disease.Disease"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_disease.Disease"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/diseases/categories": {
            "get": {
                "description": "Get unique list of disease categories for key filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diseases"
                ],
                "summary": "List disease categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/diseases/{id}": {
            "get": {
                "description": "Get full details of a specific disease",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diseases"
                ],
                "summary": "Get disease details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Disease ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_disease.Disease"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update details of an existing disease",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diseases"
                ],
                "summary": "Update disease details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Disease ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Data",
                        "name": "disease",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_disease.Disease"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_disease.Disease"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_disease.Disease": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "match_weather": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "prevention": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_disease.InfoSection"
                    }
                },
                "spread_details": {
                    "type": "string"
                },
                "symptoms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_disease.InfoSection"
                    }
                },
                "treatment": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_disease.InfoSection"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "internal_disease.InfoSection": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}